+++ server/db/db.js
@@ -1,11 +1,8 @@
-const STATES = require('../SessionStateEnum.js')
-const helpers = require('../helpers.js')
+const { ALERT_STATE, helpers } = require('brave-alert-lib')
 const SessionState = require('../SessionState.js')
 const Installation = require('../Installation.js')
 const Hub = require('../Hub.js')
 const { Pool, types } = require('pg')
-
-require('dotenv').config();
 
 const pool = new Pool({
     host: helpers.getEnvVar('PG_HOST'),
@@ -60,7 +57,7 @@ module.exports.getUnrespondedSessionWithButtonId = async function(buttonId, clie
     }
 
     const query = "SELECT * FROM sessions WHERE button_id = $1 AND state != $2 AND state != $3 AND state != $4"
-    const values = [buttonId, STATES.WAITING_FOR_CATEGORY, STATES.WAITING_FOR_DETAILS, STATES.COMPLETED]
+    const values = [buttonId, ALERT_STATE.WAITING_FOR_CATEGORY, ALERT_STATE.WAITING_FOR_DETAILS, ALERT_STATE.COMPLETED]
     const { rows } = await client.query(query, values)
    
     if(!transactionMode) {
@@ -80,7 +77,7 @@ module.exports.getMostRecentIncompleteSessionWithPhoneNumber = async function(ph
     }
     
     const query = "SELECT * FROM sessions WHERE phone_number = $1 AND state != $2 ORDER BY created_at DESC LIMIT 1"
-    const values = [phoneNumber, STATES.COMPLETED]
+    const values = [phoneNumber, ALERT_STATE.COMPLETED]
     const { rows } = await client.query(query, values)
     
     if(!transactionMode) {
@@ -168,7 +165,7 @@ module.exports.createSession = async function(installationId, buttonId, unit, ph
         client = await pool.connect()
     }
     
-    const values = [installationId, buttonId, unit, phoneNumber, STATES.STARTED, numPresses]
+    const values = [installationId, buttonId, unit, phoneNumber, ALERT_STATE.STARTED, numPresses]
     const { rows } = await client.query('INSERT INTO sessions (installation_id, button_id, unit, phone_number, state, num_presses) VALUES ($1, $2, $3, $4, $5, $6) RETURNING *', values)
    
     if(!transactionMode) {
@@ -203,9 +200,10 @@ module.exports.saveSession = async function(session, client) {
     }
 }
 
+
 module.exports.clearSessions = async function(client) {
-    if(process.env.NODE_ENV !== "test") {
-        console.log("warning - tried to clear sessions database outside of a test environment!")
+    if(!helpers.isTestEnvironment()) {
+        helpers.log("warning - tried to clear sessions database outside of a test environment!")
         return
     }
     let transactionMode = (typeof client !== 'undefined')
@@ -270,8 +268,8 @@ module.exports.createButton = async function(buttonId, installationId, unit, pho
 }
 
 module.exports.clearButtons = async function(client) {
-    if(process.env.NODE_ENV !== "test") {
-        console.log("warning - tried to clear registry database outside of a test environment!")
+    if(!helpers.isTestEnvironment()) {
+        helpers.log("warning - tried to clear registry database outside of a test environment!")
         return
     }
     let transactionMode = (typeof client !== 'undefined')
@@ -292,7 +290,7 @@ module.exports.createInstallation = async function(name, responderPhoneNumber, f
     if(!transactionMode) {
         client = await pool.connect()
     }
-    
+
     await client.query("INSERT INTO installations (name, responder_phone_number, fall_back_phone_number, incident_categories) VALUES ($1, $2, $3, $4)", [name, responderPhoneNumber, fallbackPhoneNumber, incidentCategories])
     
     if(!transactionMode) {
@@ -301,8 +299,8 @@ module.exports.createInstallation = async function(name, responderPhoneNumber, f
 }
 
 module.exports.clearInstallations = async function(client) {
-    if(process.env.NODE_ENV !== "test") {
-        console.log("warning - tried to clear installations table outside of a test environment!")
+    if(!helpers.isTestEnvironment()) {
+        helpers.log("warning - tried to clear installations table outside of a test environment!")
         return
     }
     let transactionMode = (typeof client !== 'undefined')
@@ -350,6 +348,25 @@ module.exports.getInstallationWithInstallationId = async function(installationId
     if(rows.length > 0) {
         return createInstallationFromRow(rows[0])
     }
+    return null
+}
+
+module.exports.getInstallationWithSessionId = async function(sessionId, client) {
+    let transactionMode = (typeof client !== 'undefined')
+    if(!transactionMode) {
+        client = await pool.connect()
+    }
+
+    let { rows } = await client.query("SELECT i.* FROM sessions s LEFT JOIN installations i ON s.installation_id = i.id WHERE s.id = $1", [sessionId])
+
+    if(!transactionMode) {
+        client.release()
+    }
+
+    if(rows.length > 0) {
+        return createInstallationFromRow(rows[0])
+    }
+
     return null
 }