+++ server/test/serverTest.js
@@ -4,17 +4,14 @@ const expect = chai.expect
 const { after, afterEach, beforeEach, describe, it } = require('mocha')
 const sinon = require('sinon')
 const sinonChai = require("sinon-chai")
+const { ALERT_STATE, helpers } = require('brave-alert-lib')
 
 chai.use(chaiHttp)
 chai.use(sinonChai)
 
-const STATES = require('../SessionStateEnum.js')
 const imports = require('../server.js')
 const server = imports.server
 const db = imports.db
-const twilioClient = imports.twilioClient
-
-require('dotenv').load()
 
 const sleep = (millis) => new Promise(resolve => setTimeout(resolve, millis))
 
@@ -86,7 +83,7 @@ describe('Chatbot server', () => {
             await db.clearSessions()
             await db.clearButtons()
             await db.clearInstallations()
-            console.log('\n')
+            helpers.log('\n')
         });
 
         it('should return 400 to a request with an empty body', async () => {
@@ -211,6 +208,8 @@ describe('Chatbot server', () => {
     describe('POST request: flic button press', () => {
 
         beforeEach(async function() {
+            sinon.stub(helpers, 'log')
+
             await db.clearSessions()
             await db.clearButtons()
             await db.clearInstallations()
@@ -224,7 +223,9 @@ describe('Chatbot server', () => {
             await db.clearSessions()
             await db.clearButtons()
             await db.clearInstallations()
-            console.log('\n')
+
+            helpers.log.restore()
+            helpers.log('\n')
         });
 
         it('should return 400 to a request with no headers', async () => {
@@ -241,7 +242,6 @@ describe('Chatbot server', () => {
             let responseNoSerialNumber = await chai.request(server).post('/flic_button_press').set('button-battery-level', '100').send({})
             expect(responseNoSerialNumber).to.have.status(400)
         });
-
 
         it('should return 400 to a request with an unregistered button', async () => {
             let response = await chai.request(server).post('/flic_button_press').set('button-serial-number', 'CCCC-C0C0C0').set('button-battery-level', '100').send()
@@ -381,6 +381,24 @@ describe('Chatbot server', () => {
             let buttonAgain = await db.getButtonWithSerialNumber(unit1SerialNumber)
             expect(buttonAgain.button_battery_level).to.equal(fakeBatteryLevel)
         })
+
+        it('should log a valid request when given the correct API key', async () => {
+            const buttonName = 'fakeButtonName'
+            await chai.request(server).post('/flic_button_press?apikey=testFlicApiKey').set('button-serial-number', unit1SerialNumber).set('button-name', buttonName).send()
+            expect(helpers.log).to.have.been.calledWith(`VALID api key from '${buttonName}' (${unit1SerialNumber})`)
+        })
+
+        it('should log an invalid request when given an incorrect API key', async () => {
+            const buttonName = 'fakeButtonName'
+            await chai.request(server).post('/flic_button_press?apikey=NOTtestFlicApiKey').set('button-serial-number', unit1SerialNumber).set('button-name', buttonName).send()
+            expect(helpers.log).to.have.been.calledWith(`INVALID api key from '${buttonName}' (${unit1SerialNumber})`)
+        })
+
+        it('should log a valid request when not given an API key', async () => {
+            const buttonName = 'fakeButtonName'
+            await chai.request(server).post('/flic_button_press').set('button-serial-number', unit1SerialNumber).set('button-name', buttonName).send()
+            expect(helpers.log).to.have.been.calledWith(`INVALID api key from '${buttonName}' (${unit1SerialNumber})`)
+        })
     })
 
     describe('POST request: twilio message', () => {
@@ -394,23 +412,24 @@ describe('Chatbot server', () => {
             await db.createButton(unit1UUID, installations[0].id, "1", unit1PhoneNumber, unit1SerialNumber)
             await db.createButton(unit2UUID, installations[0].id, "2", unit2PhoneNumber, unit2SerialNumber)
 
-            sinon.spy(twilioClient.messages, 'create')
+            sinon.spy(imports.braveAlerter, 'startAlertSession')
+            sinon.spy(imports.braveAlerter, 'sendSingleAlert')
         });
 
         afterEach(async function() {
-            twilioClient.messages.create.restore()
+            imports.braveAlerter.sendSingleAlert.restore()
+            imports.braveAlerter.startAlertSession.restore()
 
             await db.clearSessions()
             await db.clearButtons()
             await db.clearInstallations()
-            console.log('\n')
+            helpers.log('\n')
         });
 
         after(async function() {
-
             // wait for the staff reminder timers to finish
             await sleep(3000)
-            
+
             await db.close()
             server.close()
         })
@@ -420,7 +439,7 @@ describe('Chatbot server', () => {
             await chai.request(server).post('/').send(unit1FlicRequest_SingleClick)
             await chai.request(server).post('/').send(unit1FlicRequest_SingleClick)
 
-            expect(twilioClient.messages.create).to.be.calledOnce
+            expect(imports.braveAlerter.startAlertSession).to.be.calledOnce
         })
 
         it('should send the initial and urgent text messages after a valid double click request to /', async () => {
@@ -428,17 +447,13 @@ describe('Chatbot server', () => {
             await chai.request(server).post('/').send(unit1FlicRequest_DoubleClick)
             await chai.request(server).post('/').send(unit1FlicRequest_DoubleClick)
 
-            expect(twilioClient.messages.create).to.be.calledWith({
-                body: 'There has been a request for help from Unit 1. Please respond "Ok" when you have followed up on the call.',
-                from: sinon.match.any,
-                to: sinon.match.any,
-            })
+            expect(imports.braveAlerter.startAlertSession).to.be.calledOnce
 
-            expect(twilioClient.messages.create).to.be.calledWith({
-                body: 'This in an urgent request. The button has been pressed 2 times. Please respond "Ok" when you have followed up on the call.',
-                from: sinon.match.any, 
-                to: sinon.match.any,
-            })
+            expect(imports.braveAlerter.sendSingleAlert).to.be.calledWith(
+                sinon.match.any,
+                sinon.match.any, 
+                'This in an urgent request. The button has been pressed 2 times. Please respond "Ok" when you have followed up on the call.',
+            )
         })
 
         it('should send the initial text message after a valid single click request to /flic_button_press', async () => {
@@ -448,7 +463,7 @@ describe('Chatbot server', () => {
                 .set('button-battery-level', '100')
                 .send()
 
-            expect(twilioClient.messages.create).to.be.calledOnce
+            expect(imports.braveAlerter.startAlertSession).to.be.calledOnce
         })
 
         it('should send the initial and urgent text messages after a valid double click request to /flic_button_press', async () => {
@@ -458,17 +473,13 @@ describe('Chatbot server', () => {
                 .set('button-battery-level', '100')
                 .send()
 
-            expect(twilioClient.messages.create).to.be.calledWith({
-                body: 'There has been a request for help from Unit 1. Please respond "Ok" when you have followed up on the call.',
-                from: sinon.match.any,
-                to: sinon.match.any,
-            })
+            expect(imports.braveAlerter.startAlertSession).to.be.calledOnce
 
-            expect(twilioClient.messages.create).to.be.calledWith({
-                body: 'This in an urgent request. The button has been pressed 2 times. Please respond "Ok" when you have followed up on the call.',
-                from: sinon.match.any, 
-                to: sinon.match.any,
-            })
+            expect(imports.braveAlerter.sendSingleAlert).to.be.calledWith(
+                sinon.match.any,
+                sinon.match.any, 
+                'This in an urgent request. The button has been pressed 2 times. Please respond "Ok" when you have followed up on the call.',
+            )
         })
 
         it('should return ok to a valid request', async () => {
@@ -476,17 +487,17 @@ describe('Chatbot server', () => {
             await chai.request(server).post('/').send(unit1FlicRequest_SingleClick)
             await chai.request(server).post('/').send(unit1FlicRequest_SingleClick)
 
-            let response = await chai.request(server).post('/message').send(twilioMessageUnit1_InitialStaffResponse);
+            let response = await chai.request(server).post('/alert/sms').send(twilioMessageUnit1_InitialStaffResponse);
             expect(response).to.have.status(200);
         });
 
         it('should return 400 to a request with incomplete data', async () => {
-            let response = await chai.request(server).post('/message').send({'Body': 'hi'});
+            let response = await chai.request(server).post('/alert/sms').send({'Body': 'hi'});
             expect(response).to.have.status(400);
         });
 
         it('should return 400 to a request from an invalid phone number', async () => {
-            let response = await chai.request(server).post('/message').send({'Body': 'hi', 'From': '+16664206969'});
+            let response = await chai.request(server).post('/alert/sms').send({'Body': 'hi', 'From': '+16664206969'});
             expect(response).to.have.status(400);
         });
 
@@ -497,28 +508,28 @@ describe('Chatbot server', () => {
 
             let sessions = await db.getAllSessionsWithButtonId(unit1UUID)
             expect(sessions.length).to.equal(1)
-            expect(sessions[0].state, 'state after initial button press').to.deep.equal(STATES.STARTED);
+            expect(sessions[0].state, 'state after initial button press').to.deep.equal(ALERT_STATE.STARTED);
 
-            let response = await chai.request(server).post('/message').send(twilioMessageUnit1_InitialStaffResponse);
+            let response = await chai.request(server).post('/alert/sms').send(twilioMessageUnit1_InitialStaffResponse);
             expect(response).to.have.status(200);
 
             sessions = await db.getAllSessionsWithButtonId(unit1UUID)
             expect(sessions.length).to.equal(1)
-            expect(sessions[0].state, 'state after initial staff response').to.deep.equal(STATES.WAITING_FOR_CATEGORY);
+            expect(sessions[0].state, 'state after initial staff response').to.deep.equal(ALERT_STATE.WAITING_FOR_CATEGORY);
 
-            response = await chai.request(server).post('/message').send(twilioMessageUnit1_IncidentCategoryResponse)
+            response = await chai.request(server).post('/alert/sms').send(twilioMessageUnit1_IncidentCategoryResponse)
             expect(response).to.have.status(200)
  
             sessions = await db.getAllSessionsWithButtonId(unit1UUID)
             expect(sessions.length).to.equal(1)
-            expect(sessions[0].state, 'state after staff have categorized the incident').to.deep.equal(STATES.WAITING_FOR_DETAILS)
+            expect(sessions[0].state, 'state after staff have categorized the incident').to.deep.equal(ALERT_STATE.WAITING_FOR_DETAILS)
 
-            response = await chai.request(server).post('/message').send(twilioMessageUnit1_IncidentNotesResponse)
+            response = await chai.request(server).post('/alert/sms').send(twilioMessageUnit1_IncidentNotesResponse)
             expect(response).to.have.status(200)
  
             sessions = await db.getAllSessionsWithButtonId(unit1UUID)
             expect(sessions.length).to.equal(1)
-            expect(sessions[0].state, 'state after staff have provided incident notes').to.deep.equal(STATES.COMPLETED) 
+            expect(sessions[0].state, 'state after staff have provided incident notes').to.deep.equal(ALERT_STATE.COMPLETED) 
 
             // now start a new session for a different unit
             // Power Automate always sends two requests
@@ -527,7 +538,7 @@ describe('Chatbot server', () => {
 
             sessions = await db.getAllSessionsWithButtonId(unit2UUID)
             expect(sessions.length).to.equal(1)
-            expect(sessions[0].state, 'state after new button press from a different unit').to.deep.equal(STATES.STARTED)
+            expect(sessions[0].state, 'state after new button press from a different unit').to.deep.equal(ALERT_STATE.STARTED)
             expect(sessions[0].buttonId).to.deep.equal(unit2UUID)
             expect(sessions[0].unit).to.deep.equal('2')
             expect(sessions[0].numPresses).to.deep.equal(1)
@@ -538,11 +549,11 @@ describe('Chatbot server', () => {
             let response = await chai.request(server).post('/').send(unit1FlicRequest_SingleClick)
             response = await chai.request(server).post('/').send(unit1FlicRequest_SingleClick)
 
-            expect(response).to.have.status(200);
+            expect(response).to.have.status(200)
             await sleep(4000)
             let sessions = await db.getAllSessionsWithButtonId(unit1UUID)
             expect(sessions.length).to.equal(1)
-            expect(sessions[0].state, 'state after reminder timeout has elapsed').to.deep.equal(STATES.WAITING_FOR_REPLY);
+            expect(sessions[0].state, 'state after reminder timeout has elapsed').to.deep.equal(ALERT_STATE.WAITING_FOR_REPLY);
             expect(sessions[0].fallBackAlertTwilioStatus).to.not.be.null
             expect(sessions[0].fallBackAlertTwilioStatus).to.not.equal('failed')
             expect(sessions[0].fallBackAlertTwilioStatus).to.not.equal('undelivered')